Fix object list that need READ permission
Read permission cannot be granted over V$xxx objects  since they are synonyms and not tables or views and ORA-02030 is thrown. Change them to the underlying view names instead.

Also remark that READ permission is enough and should be used instead of SELECT permission if Oracle version is 12 or greater.